Add Print character info function
Program not finds a character (it also subs spaces with '+' so it works on Tibia website) and then using XPath, it prints the character information table with the  relevant info from Tibia characters. Optional fields such as former names or former world are also considered and printed if they exist.